+++ atmega_src/test_atmega.h
@@ -1,0 +1,6 @@
+#ifndef ATMEGA_SRC_H_
+#define ATMEGA_SRC_H_
+
+void run_test();
+
+#endif
