--- atmega_src/travis_ci_test_example.h
@@ -1,6 +1,0 @@
-#ifndef MAIN_H_
-#define MAIN_H_
-
-int run_tests_default();
-
-#endif /* MAIN_H_ */
